% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callLength.R
\name{callLength}
\alias{callLength}
\title{call fragment length}
\usage{
callLength(
  fragment_obj,
  fragment_type = "default",
  isize_min = 1L,
  isize_max = 1000L,
  ...
)
}
\arguments{
\item{fragment_type}{String; "default", "mut_vs_ref", "mut_vs_outer", "mut_vs_ref_norm", "mut_vs_outer_norm"}

\item{...}{}
}
\value{
a tibble
}
\description{
call fragment length
}
