% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{check_mutfile_columns}
\alias{check_mutfile_columns}
\title{Internal Functions for processing mutational fragment-level information}
\usage{
check_mutfile_columns(df)
}
\description{
Internal Functions for processing mutational fragment-level information
}
