% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{clipReadSeq}
\alias{clipReadSeq}
\title{Remove soft-clipped bases from read sequences}
\usage{
clipReadSeq(vdf = NULL)
}
\arguments{
\item{vdf}{}
}
\value{
a dataframe
}
\description{
Remove soft-clipped bases from read sequences
}
