% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{mismatchPosition}
\alias{mismatchPosition}
\title{Obtain internal position of mismatch within read}
\usage{
mismatchPosition(
  read_id,
  chr_start,
  start_pos,
  read_strand,
  nm_tag,
  md_tag,
  cigar_string,
  clip_seq,
  which_locus
)
}
\arguments{
\item{which_locus}{}
}
\value{
a dataframe
}
\description{
Obtain internal position of mismatch within read
}
