% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrinucleotide.R
\name{plotTrinucleotide}
\alias{plotTrinucleotide}
\title{Plot Trinucleotide Data}
\usage{
plotTrinucleotide(
  trinuc_df,
  exclude_if_type_present = NULL,
  retain_if_type_present = NULL,
  remove_type = NULL,
  normalize_counts = TRUE,
  show_overlap_type = TRUE,
  min_y_val = 0,
  max_y_val = 0.5,
  plot_title = "Trinucleotide Profile",
  y_axis_title = "Percentage of Single Base Substitutions",
  draw_x_axis_labels = TRUE,
  draw_y_axis_labels = TRUE,
  draw_y_axis_title = TRUE,
  save_plot = FALSE,
  output_file = "./trinucleotide_profile.pdf"
)
}
\arguments{
\item{trinuc_df}{DataFrame containing trinucleotide data.}

\item{exclude_if_type_present}{Vector of strings specifying column names whose non-zero presence should trigger exclusion of rows.}

\item{retain_if_type_present}{Vector of strings specifying column names whose non-zero presence is necessary to retain rows.}

\item{remove_type}{Vector of strings specifying column names where all values should be set to zero.}

\item{normalize_counts}{Logical; if TRUE, normalizes the 'value' column to sum to 1.}

\item{show_overlap_type}{Logical; if TRUE, show read-pair overlap types.}

\item{min_y_val}{Numeric; minimum y-axis value for the plots.}

\item{max_y_val}{Numeric; maximum y-axis value for the plots.}

\item{plot_title}{String; title for the plots.}

\item{y_axis_title}{String; title for the y-axis.}

\item{draw_x_axis_labels}{Logical; whether to draw x-axis labels.}

\item{draw_y_axis_labels}{Logical; whether to draw y-axis labels.}

\item{draw_y_axis_title}{Logical; whether to display a title for the y-axis.}

\item{save_plot}{Logical; whether to save the object as a pdf file.}

\item{output_file}{String; name and path of output pdf file.}

\item{overlap_type}{Logical; whether to stratify SBS trinucleotides by read pair overlap type.}
}
\value{
A trinucleotide SBS plot object and an optional pdf file.
}
\description{
This function processes and plots trinucleotide data. It first applies specified
filters and transformations to the data and then generates a visual representation
of the results. The function handles data normalization, exclusion, and retention
based on provided column names, and it creates detailed plots with options for
customization of plot aesthetics.
}
\examples{
\dontrun{
 plotTrinucleotide(trinuc_df)
}

}
