% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_bam_file}
\alias{process_bam_file}
\title{Function to process BAM files based on mutational or general alignment data}
\usage{
process_bam_file(
  bamfile = bamfile,
  mutation_file = mutation_file,
  mut_fragments_only = mut_fragments_only,
  use_names = use_names,
  chromosome_to_keep = chromosome_to_keep,
  strand_mode = strand_mode,
  genome_name = genome_name,
  call_mutations = call_mutations,
  galp_flag = galp_flag,
  galp_what = galp_what,
  galp_tag = galp_tag,
  galp_mapqFilter = galp_mapqFilter
)
}
\description{
Function to process BAM files based on mutational or general alignment data
}
