% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_mutation_fragments}
\alias{process_mutation_fragments}
\title{Process and curate mutational fragment-level information}
\usage{
process_mutation_fragments(
  bamfile = bamfile,
  use_names = use_names,
  chromosome_to_keep = chromosome_to_keep,
  strand_mode = strand_mode,
  genome = genome,
  galp_flag = galp_flag,
  galp_what = galp_what,
  galp_tag = galp_tag,
  galp_mapqFilter = galp_mapqFilter,
  mutation_file = mutation_file,
  call_mutations = call_mutations,
  pileup_max_depth = pileup_max_depth,
  pileup_min_base_quality = pileup_min_base_quality,
  pileup_min_mapq = pileup_min_mapq,
  pileup_min_nucleotide_depth = pileup_min_nucleotide_depth,
  pileup_min_minor_allele_depth = pileup_min_minor_allele_depth,
  pileupYieldSize = pileupYieldSize,
  frag = frag
)
}
\description{
Process and curate mutational fragment-level information
}
