% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{processMismatches}
\alias{processMismatches}
\title{Process Mismatch data}
\usage{
processMismatches(
  which_loci,
  loci_df,
  bamfile,
  galp_flag,
  galp_mapqFilter,
  chromosome_to_keep
)
}
\arguments{
\item{galp_mapqFilter}{}
}
\value{
a Dataframe of processes mismatches
}
\description{
Process Mismatch data
}
